Add files via upload
PROBLEM DEFINITION 
Manually maintaining the records for every product is an extremely time-consuming task that 
is also quite tough. Managing the details is not recommended because if any issues happen, 
such as missing records that were stored offline, several issues will arise and it would be 
impossible to recover the data. 
The problem faced by the company is they do not have any systematic system to record and 
keep their stock data. It is difficult for the admin to record the stock data quickly and safely 
because they only keep it in the logbook and not properly organized.